requirements.txt
aniso8601==8.0.0
click==7.1.2
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-RESTful==0.3.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.0
Pillow==7.2.0
scipy==1.5.1
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
keras
tensorflow
opencv-python==4.3.0.36
flask-cors